Minor fixes and relocation of LogNormal Prior to modeling
