+++ gammapy/modeling/models/tests/test_prior.py
@@ -1,18 +1,20 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-import pytest
-import numpy as np
-from numpy.testing import assert_allclose
 import astropy.units as u
+import numpy as np
+import pytest
+from numpy.testing import assert_allclose
+
 from gammapy.modeling import Parameter
 from gammapy.modeling.models import (
     PRIOR_REGISTRY,
     GaussianPrior,
     GeneralizedGaussianPrior,
+    LogNormalNuisancePrior,
+    LogUniformPrior,
     Model,
     Models,
     SkyModel,
     UniformPrior,
-    LogUniformPrior,
 )
 from gammapy.utils.testing import assert_quantity_allclose
 
@@ -200,3 +202,57 @@ def test_factor_min_max_use_synced_bounds():
     p.prior = UniformPrior(min=0.0, max=1e-10)
     assert_allclose(p.factor_min, 0.0)
     assert_allclose(p.factor_max, 1e-10 / 1e-12)
+
+
+def test_lognormal_prior_evaluate_at_center():
+    """Prior must be zero at the observed value."""
+    prior = LogNormalNuisancePrior(log10_obs=19.3, sigma_stat=0.1)
+    test_val = Parameter(name="test_val", value=19.3)
+    assert_allclose(prior(test_val), 0.0, atol=1e-10)
+
+
+def test_lognormal_prior_evaluate_one_sigma():
+    """Prior must be 1.0 exactly one sigma away."""
+    prior = LogNormalNuisancePrior(log10_obs=19.3, sigma_stat=0.1)
+    assert_allclose(prior(Parameter(name="test_val", value=19.4)), 1.0, rtol=1e-5)
+    assert_allclose(prior(Parameter(name="test_val", value=19.2)), 1.0, rtol=1e-5)
+
+
+def test_lognormal_prior_sigma_in_quadrature():
+    """sigma_total must be computed as sqrt(stat² + syst²)."""
+    prior = LogNormalNuisancePrior(log10_obs=19.3, sigma_stat=0.3, sigma_syst=0.4)
+    assert_allclose(prior.sigma_total.value, 0.5, rtol=1e-6)
+
+
+def test_lognormal_prior_only_syst():
+    prior = LogNormalNuisancePrior(log10_obs=19.3, sigma_syst=0.2)
+    assert_allclose(prior.sigma_total.value, 0.2, rtol=1e-6)
+    assert_allclose(prior(Parameter(name="test_val", value=19.5)), 1.0, rtol=1e-5)
+
+
+def test_lognormal_prior_invalid_sigma_stat():
+    with pytest.raises((TypeError, ValueError)):
+        LogNormalNuisancePrior(log10_obs=19.3, sigma_stat=-0.1)
+
+    with pytest.raises((TypeError, ValueError)):
+        LogNormalNuisancePrior(log10_obs=19.3, sigma_stat="bad")
+
+
+def test_lognormal_prior_invalid_sigma_syst():
+    with pytest.raises((TypeError, ValueError)):
+        LogNormalNuisancePrior(log10_obs=19.3, sigma_syst="bad")
+
+    with pytest.raises((TypeError, ValueError)):
+        LogNormalNuisancePrior(log10_obs=19.3, sigma_syst=-0.1)
+
+
+def test_validate_sigma_none_returns_zero():
+    from gammapy.modeling.models.prior import _validate_sigma
+
+    assert_allclose(_validate_sigma("test", None), 0.0)
+
+
+def test_lognormal_prior_sigma_none_converts_to_zero():
+    prior = LogNormalNuisancePrior(log10_obs=19.3, sigma_stat=None, sigma_syst=None)
+    assert_allclose(prior.sigma_stat, 0.0)
+    assert_allclose(prior.sigma_syst, 0.0)
